requirements.txt
torch=0.4.1
torchvision=0.2.1
graphviz==0.9
pytorchviz==0.0.1
# git+https://github.com/szagoruyko/pytorchviz
hypothesis==3.70.3
pytest==3.8.0